% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-sets.R
\name{are_set_equal}
\alias{are_set_equal}
\alias{is_set_equal}
\alias{is_subset}
\alias{is_superset}
\alias{are_disjoint_sets}
\alias{are_intersecting_sets}
\alias{assert_is_set_equal}
\alias{assert_is_subset}
\alias{assert_is_superset}
\alias{assert_are_set_equal}
\alias{assert_are_disjoint_sets}
\alias{assert_are_intersecting_sets}
\title{Set comparisons}
\description{
See \code{\link[assertive.sets]{are_set_equal}}.
}
