% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-connection.R
\name{as.character.file}
\alias{as.character.file}
\title{Convert file connections to strings}
\usage{
\method{as.character}{file}(x, ...)
}
\arguments{
\item{x}{A file connection.}

\item{...}{Not currently used.}
}
\value{
A string containing the target location of the file connection.
}
\description{
\code{as.character} method for file connections.
}
\examples{
rprofile <- file.path(R.home("etc"), "Rprofile.site")
fcon <- file(rprofile)
assert_all_are_true(identical(as.character(fcon), rprofile))
close(fcon)
}
\seealso{
\code{\link[base]{file}}, \code{\link[base]{summary.connection}},
\code{\link[base]{as.character}}
}

