% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{DIM}
\alias{DIM}
\title{Get the dimensions of an object}
\usage{
DIM(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A integer vector of non-negative values.
}
\description{
Get the dimensions of an object, retuning the length if that object has no
\code{dim} attribute.
}
\examples{
# For data frames and matrices, DIM is the same as dim.
DIM(sleep)
# For vectors (and other objects without a dim attribute), DIM is the
# same as length.
DIM(1:10)
DIM(list(x = 1:10))
}
\seealso{
\code{\link[base]{NROW}}, \code{\link[base]{dim}}
}

