% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{dont_stop}
\alias{dont_stop}
\title{Run code without stopping}
\usage{
dont_stop(...)
}
\arguments{
\item{...}{Passed to \code{tryCatch}.}
}
\value{
The expression that was passed in is run.
}
\description{
Runs code without stopping, warnings and errors are only printed.
}
\note{
This function is dangerous, since it overrides warnings and errors.
Its intended use is for documenting examples of errors.
}
\examples{
dont_stop(warning("!!!"))
dont_stop(stop("!!!"))
f <- function() g()
g <- function() stop("!!!")
dont_stop(f())
}

