\name{get_name_in_parent}
\alias{get_name_in_parent}
\title{Get the name of a variable in the parent frame.}
\usage{
  get_name_in_parent(x)
}
\arguments{
  \item{x}{Variable to get the name of.}
}
\value{
  A string giving the name of the input in the parent
  frame.
}
\description{
  Gets the name of the input in the parent frame.
}

