% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{has_any_attributes}
\alias{has_any_attributes}
\alias{has_no_attributes}
\title{Does the input have any attributes?}
\description{
See \code{\link[assertive.properties]{has_any_attributes}}.
}

