% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{has_attributes}
\alias{assert_has_all_attributes}
\alias{assert_has_any_attributes}
\alias{has_attributes}
\title{Does the input have the specified attributes?}
\description{
See \code{\link[assertive.properties]{has_attributes}}.
}

