% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{is_atomic}
\alias{assert_is_atomic}
\alias{assert_is_recursive}
\alias{assert_is_vector}
\alias{is_atomic}
\alias{is_recursive}
\alias{is_vector}
\title{Is the input atomic/recursive/vector?}
\description{
See \code{\link[assertive.properties]{is_atomic}}.
}

