% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_batch_mode}
\alias{assert_is_batch_mode}
\alias{assert_is_interactive}
\alias{assert_is_r_slave}
\alias{assert_is_slave_r}
\alias{is_batch_mode}
\alias{is_interactive}
\alias{is_r_slave}
\alias{is_slave_r}
\title{How is R running?}
\description{
See \code{\link[assertive.reflection]{is_batch_mode}}.
}

