% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_cas_number}
\alias{assert_all_are_cas_numbers}
\alias{assert_any_are_cas_numbers}
\alias{is_cas_number}
\title{Does the character vector contain CAS registry numbers?}
\description{
See \code{\link[assertive.data]{is_cas_number}}.
}

