% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_closure_function}
\alias{assert_is_builtin_function}
\alias{assert_is_closure_function}
\alias{assert_is_special_function}
\alias{is_builtin_function}
\alias{is_closure_function}
\alias{is_special_function}
\title{Is the input a closure, builtin or special function?}
\description{
See \code{\link[assertive.types]{is_closure_function}}.
}

