% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_complex}
\alias{assert_is_a_complex}
\alias{assert_is_complex}
\alias{is_a_complex}
\alias{is_complex}
\title{Is the input complex?}
\description{
See \code{\link[assertive.types]{is_complex}}.
}

