% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assert-is-reflection.R, R/is-reflection.R
\name{assert_is_current_r}
\alias{assert_is_current_r}
\alias{is_current_r}
\title{Is this version of R up to date?}
\usage{
assert_is_current_r()

is_current_r(cran = getOption("repos", c(CRAN =
  "http://cran.r-project.org"))["CRAN"])
}
\arguments{
\item{cran}{A string giving the URL of the CRAN repository to check.}
}
\value{
An object of class \code{R_system_version} giving the current release
version of R.
}
\description{
Check if this version of R is as new as the current release version of R.
}
\note{
Development versions of R can have versions higher than the current
release version of R.  For convenience, these will return \code{TRUE}.
}
\examples{
is_current_r()
}

