% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_date}
\alias{assert_is_date}
\alias{assert_is_posixct}
\alias{assert_is_posixlt}
\alias{is_date}
\alias{is_posixct}
\alias{is_posixlt}
\title{Is the input a date?}
\description{
See \code{\link[assertive.types]{is_date}}.
}

