% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_debugged}
\alias{assert_is_debugged}
\alias{is_debugged}
\title{Is the input function being debugged?
 
See \code{\link[assertive.code]{is_debugged}}.}
\description{
Is the input function being debugged?
 
See \code{\link[assertive.code]{is_debugged}}.
}

