% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_email_address}
\alias{assert_all_are_email_addresses}
\alias{assert_any_are_email_addresses}
\alias{is_email_address}
\title{Does the character vector contain email addresses?}
\description{
See \code{\link[assertive.data]{is_email_address}}.
}

