% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{is_empty}
\alias{assert_has_elements}
\alias{assert_is_empty}
\alias{assert_is_non_empty}
\alias{assert_is_non_scalar}
\alias{assert_is_of_dimension}
\alias{assert_is_of_length}
\alias{assert_is_scalar}
\alias{has_elements}
\alias{is_empty}
\alias{is_non_empty}
\alias{is_non_scalar}
\alias{is_of_dimension}
\alias{is_of_length}
\alias{is_scalar}
\title{Is the input empty/scalar?}
\description{
See \code{\link[assertive.properties]{is_empty}}.
}

