% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-numbers.R
\name{is_equal_to}
\alias{is_equal_to}
\alias{is_not_equal_to}
\alias{is_greater_than}
\alias{is_greater_than_or_equal_to}
\alias{is_less_than}
\alias{is_less_than_or_equal_to}
\alias{assert_all_are_equal_to}
\alias{assert_any_are_equal_to}
\alias{assert_all_are_not_equal_to}
\alias{assert_any_are_not_equal_to}
\alias{assert_all_are_greater_than}
\alias{assert_any_are_greater_than}
\alias{assert_all_are_greater_than_or_equal_to}
\alias{assert_any_are_greater_than_or_equal_to}
\alias{assert_all_are_less_than}
\alias{assert_any_are_less_than}
\alias{assert_all_are_less_than_or_equal_to}
\alias{assert_any_are_less_than_or_equal_to}
\title{How does the input relate to a value?}
\description{
See \code{\link[assertive.numbers]{is_equal_to}}.
}
