% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_error_free}
\alias{is_error_free}
\title{Does the code run without throwing an error?
 
See \code{\link[assertive.code]{is_error_free}}.}
\description{
Does the code run without throwing an error?
 
See \code{\link[assertive.code]{is_error_free}}.
}
