% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-files.R
\name{is_executable_file}
\alias{assert_all_are_ex_files}
\alias{assert_all_are_executable_files}
\alias{assert_all_are_readable_files}
\alias{assert_all_are_writable_files}
\alias{assert_any_are_ex_files}
\alias{assert_any_are_executable_files}
\alias{assert_any_are_readable_files}
\alias{assert_any_are_writable_files}
\alias{is_ex_file}
\alias{is_executable_file}
\alias{is_readable_file}
\alias{is_writable_file}
\title{Is the file accessible?}
\description{
See \code{\link[assertive.files]{is_executable_file}}.
}

