% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-files.R
\name{is_existing_file}
\alias{assert_all_are_existing_files}
\alias{assert_any_are_existing_files}
\alias{is_existing_file}
\title{Does the file exist?}
\description{
See \code{\link[assertive.files]{is_existing_file}}.
}

