% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_formula}
\alias{assert_is_formula}
\alias{assert_is_one_sided_formula}
\alias{assert_is_two_sided_formula}
\alias{is_formula}
\alias{is_one_sided_formula}
\alias{is_two_sided_formula}
\title{Is the input a formula?}
\description{
See \code{\link[assertive.types]{is_formula}}.
}

