% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_hex_color}
\alias{assert_all_are_hex_colors}
\alias{assert_all_are_hex_colours}
\alias{assert_any_are_hex_colors}
\alias{assert_any_are_hex_colours}
\alias{is_hex_color}
\alias{is_hex_colour}
\title{Does the character vector contain hex colors?}
\description{
See \code{\link[assertive.data]{is_hex_color}}.
}

