% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-matrices.R
\name{is_identity_matrix}
\alias{assert_is_identity_matrix}
\alias{is_identity_matrix}
\title{Is the input an identity matrix?}
\description{
See \code{\link[assertive.matrices]{is_identity_matrix}}.
}

