% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_if_condition}
\alias{assert_is_if_condition}
\alias{is_if_condition}
\title{Is suitable to be used as an if condition}
\description{
See \code{\link[assertive.code]{is_if_condition}}.
}

