% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_inherited_from}
\alias{assert_is_inherited_from}
\alias{is_inherited_from}
\title{Does the object inherit from some class?}
\description{
See \code{\link[assertive.types]{is_inherited_from}}.
}

