% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_internal_function}
\alias{assert_is_internal_function}
\alias{is_internal_function}
\title{Is the input an internal function?}
\description{
See \code{\link[assertive.types]{is_closure_function}}.
}

