% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_isbn_code}
\alias{assert_all_are_isbn_codes}
\alias{assert_any_are_isbn_codes}
\alias{is_isbn10_code}
\alias{is_isbn13_code}
\alias{is_isbn_code}
\title{Does the character vector contain ISBN book codes?}
\description{
See \code{\link[assertive.data]{is_isbn_code}}.
}

