% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-matrices.R
\name{is_lower_triangular_matrix}
\alias{assert_is_lower_triangular_matrix}
\alias{assert_is_upper_triangular_matrix}
\alias{is_lower_triangular_matrix}
\alias{is_upper_triangular_matrix}
\title{Is the matrix upper/lower triangular?}
\description{
See \code{\link[assertive.matrices]{is_lower_triangular_matrix}}.
}

