% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-strings.R
\name{is_matching_fixed}
\alias{is_matching_fixed}
\alias{is_not_matching_fixed}
\alias{is_matching_regex}
\alias{is_not_matching_regex}
\alias{assert_all_are_matching_fixed}
\alias{assert_any_are_matching_fixed}
\alias{assert_all_are_not_matching_fixed}
\alias{assert_any_are_not_matching_fixed}
\alias{assert_all_are_matching_regex}
\alias{assert_any_are_matching_regex}
\alias{assert_all_are_not_matching_regex}
\alias{assert_any_are_not_matching_regex}
\title{Does the string match a pattern?
 
See \code{\link[assertive.strings]{is_matching_fixed}}.}
\description{
Does the string match a pattern?
 
See \code{\link[assertive.strings]{is_matching_fixed}}.
}
