% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{is_monotonic_increasing}
\alias{is_monotonic_increasing}
\alias{is_monotonic}
\alias{is_monotonic_decreasing}
\alias{assert_is_monotonic_increasing}
\alias{assert_is_monotonic_decreasing}
\title{Is the vector monotonically increasing or decreasing?}
\description{
See \code{\link[assertive.properties]{is_monotonic_increasing}}.
}
