% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-numbers.R
\name{is_nan}
\alias{assert_all_are_nan}
\alias{assert_all_are_not_nan}
\alias{assert_any_are_nan}
\alias{assert_any_are_not_nan}
\alias{is_nan}
\alias{is_not_nan}
\title{Is the input (not) NaN?}
\description{
See \code{\link[assertive.numbers]{is_nan}}.
}

