% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_numeric}
\alias{assert_is_a_number}
\alias{assert_is_numeric}
\alias{is_a_number}
\alias{is_numeric}
\title{Is the input numeric?}
\description{
See \code{\link[assertive.types]{is_numeric}}.
}

