% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-strings.R
\name{is_numeric_string}
\alias{assert_all_are_numeric_strings}
\alias{assert_any_are_numeric_strings}
\alias{is_numeric_string}
\title{Does the string contain a number?}
\description{
See \code{\link[assertive.strings]{is_numeric_string}}.
}

