% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_r_current}
\alias{assert_is_current_r}
\alias{assert_is_r_current}
\alias{is_r_current}
\title{Is this version of R up to date?}
\description{
See \code{\link[assertive.reflection]{is_r_current}}.
}

