% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_rstudio_current}
\alias{assert_is_rstudio_current}
\alias{is_rstudio_current}
\title{Is RStudio the current version?}
\description{
See \code{\link[assertive.reflection]{is_rstudio_current}}.
}

