% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_rstudio_desktop}
\alias{assert_is_rstudio_desktop}
\alias{assert_is_rstudio_server}
\alias{is_rstudio_desktop}
\alias{is_rstudio_server}
\title{Is RStudio running in desktop or server mode?}
\description{
See \code{\link[assertive.reflection]{is_rstudio_desktop}}.
}

