% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_s3_generic}
\alias{is_s3_generic}
\alias{is_s3_method}
\alias{is_s3_primitive_generic}
\alias{is_s3_group_generic}
\alias{is_s3_internal_generic}
\alias{is_s4_group_generic}
\alias{assert_is_s3_generic}
\alias{assert_is_s3_method}
\alias{assert_is_s3_primitive_generic}
\alias{assert_is_s3_group_generic}
\alias{assert_is_s3_internal_generic}
\alias{assert_is_s4_group_generic}
\title{Is the input an S3 generic or method?}
\description{
See \code{\link[assertive.types]{is_s3_generic}}.
}
