% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_s4}
\alias{assert_is_S4}
\alias{assert_is_ref_class_generator}
\alias{assert_is_ref_class_object}
\alias{assert_is_s4}
\alias{is_S4}
\alias{is_ref_class_generator}
\alias{is_ref_class_object}
\alias{is_s4}
\title{Is the input an S4 object?}
\description{
See \code{\link[assertive.types]{is_s4}}.
}

