% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-sets.R
\name{is_set_equal}
\alias{assert_are_set_equal}
\alias{assert_is_set_equal}
\alias{assert_is_subset}
\alias{assert_is_superset}
\alias{is_set_equal}
\alias{is_subset}
\alias{is_superset}
\title{Set comparisons}
\description{
See \code{\link[assertive.sets]{is_set_equal}}.
}

