% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-strings.R
\name{is_single_character}
\alias{assert_all_are_single_characters}
\alias{assert_any_are_single_characters}
\alias{is_single_character}
\title{Is the input a single character?}
\description{
See \code{\link[assertive.strings]{is_single_character}}.
}

