% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-matrices.R
\name{is_square_matrix}
\alias{assert_is_square_matrix}
\alias{is_square_matrix}
\title{Is the input a square matrix?}
\description{
See \code{\link[assertive.matrices]{is_square_matrix}}.
}

