% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_tbl}
\alias{assert_is_tbl}
\alias{assert_is_tbl_cube}
\alias{assert_is_tbl_df}
\alias{assert_is_tbl_dt}
\alias{is_tbl}
\alias{is_tbl_cube}
\alias{is_tbl_df}
\alias{is_tbl_dt}
\title{Is the input a tbl?}
\description{
See \code{\link[assertive.types]{is_tbl}}.
}

