% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_try_error}
\alias{is_try_error}
\alias{is_simple_error}
\alias{is_error}
\alias{is_simple_warning}
\alias{is_warning}
\alias{is_simple_message}
\alias{is_message}
\alias{is_condition}
\alias{assert_is_try_error}
\alias{assert_is_simple_error}
\alias{assert_is_error}
\alias{assert_is_simple_warning}
\alias{assert_is_warning}
\alias{assert_is_simple_message}
\alias{assert_is_message}
\alias{assert_is_condition}
\title{Is the input a condition?}
\description{
See \code{\link[assertive.types]{is_try_error}}.
}
