% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-data-us.R
\name{is_us_social_security_number}
\alias{assert_all_are_us_social_security_numbers}
\alias{assert_any_are_us_social_security_numbers}
\alias{is_us_social_security_number}
\title{Is the string a valid US SSN?}
\description{
See \code{\link[assertive.data.us]{is_us_social_security_number}}.
}

