% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_valid_variable_name}
\alias{is_valid_variable_name}
\alias{assert_all_are_valid_variable_names}
\alias{assert_any_are_valid_variable_names}
\title{Is the string a valid variable name?
 
See \code{\link[assertive.code]{is_valid_variable_name}}.}
\description{
Is the string a valid variable name?
 
See \code{\link[assertive.code]{is_valid_variable_name}}.
}
