% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_xxx_for_decimal_point}
\alias{assert_is_comma_for_decimal_point}
\alias{assert_is_period_for_decimal_point}
\alias{is_comma_for_decimal_point}
\alias{is_period_for_decimal_point}
\alias{is_xxx_for_decimal_point}
\title{What does the current locale specify for the decimal point?}
\description{
See \code{\link[assertive.reflection]{is_xxx_for_decimal_point}}.
}

