% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assert-is-in-range.R
\name{make_range_string}
\alias{make_range_string}
\title{Make a range string}
\usage{
make_range_string(lower, upper, lower_is_strict, upper_is_strict)
}
\arguments{
\item{lower}{A number for the lower bound.}

\item{upper}{A number for the upper bound.}

\item{lower_is_strict}{Should the lower bound be included?}

\item{upper_is_strict}{Should the upper bound be included?}
}
\value{
A string denoting the range.
}
\description{
Makes a range string using mathematical notation.
}
\note{
Not vectorized across the \code{lower_is_strict} and
\code{upper_is_strict} args for speed.
}
\examples{
\donttest{
make_range_string(-1.2345, 6.7890, TRUE, FALSE)
}
}

