% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_dots_with_list}
\alias{merge_dots_with_list}
\title{Merge ellipsis args with a list.}
\usage{
merge_dots_with_list(..., l = list())
}
\arguments{
\item{...}{Some inputs.}

\item{l}{A list.}
}
\value{
A list containing the merged inputs.
}
\description{
Merges variable length ellipsis arguments to a function with a list argument.
}
\note{
If any arguments are present in both the \code{...} and \code{l}
arguments, the \code{...} version takes preference, and a warning is thrown.
}
\examples{
merge_dots_with_list(
  foo = 1,
  bar = 2,
  baz = 3,
  l = list(foo = 4, baz = 5, quux = 6)
)
}
\seealso{
\code{\link{merge.list}}, \code{\link[base]{merge}}
}

