\name{modal_value}
\alias{modal_value}
\title{The most common value in a vector.}
\usage{
  modal_value(x)
}
\arguments{
  \item{x}{vector to find the modal value of.}
}
\value{
  The modal value of \code{x}.
}
\description{
  The modal value of a vector.
}
\note{
  Probably very inefficient; not suitable for general use.
}

