% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-base.R, R/exports-properties.R
\name{n_elements}
\alias{n_elements}
\title{Get the number of elements}
\description{
See \code{\link[assertive.properties]{n_elements}}.

See \code{\link[assertive.properties]{n_elements}}.
}

