% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-reflection.R
\name{not_this_build}
\alias{not_this_build}
\title{Failure for bad build}
\usage{
not_this_build(status)
}
\arguments{
\item{status}{A string giving the name of the build status that was desired.}
}
\value{
A string showing the actual build status.
}
\description{
Wrapper to \code{false} for failure messages when the OS is not as
expected.
}
\examples{
\donttest{
assertive:::not_this_build("stable")
assertive:::not_this_build("development")
}
}
\seealso{
\code{\link[base]{.Platform}} and \code{\link[base]{Sys.info}}
}

