% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-reflection.R
\name{not_this_os}
\alias{not_this_os}
\title{Failure for bad OS}
\usage{
not_this_os(os)
}
\arguments{
\item{os}{A string giving the name of the OS that was desired.}
}
\value{
A string showing the results of \code{.Platform$OS} and
\code{Sys.info()['sysname']}.
}
\description{
Wrapper to \code{false} for failure messages when the OS is not as
expected.
}
\examples{
\donttest{
assertive:::not_this_os("Windows")
assertive:::not_this_os("BSD-based")
}
}
\seealso{
\code{\link[base]{.Platform}} and \code{\link[base]{Sys.info}}
}

