% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{print_and_capture}
\alias{print_and_capture}
\title{Print a variable and capture the output}
\usage{
print_and_capture(x)
}
\arguments{
\item{x}{A variable.}
}
\value{
A string.
}
\description{
Prints a variable and captures the output, collapsing the value to a single
string.
}
\examples{
\dontrun{
# This is useful for including data frames in warnings or errors
message("This is the CO2 dataset:\\n", print_and_capture(CO2))
}
}
\seealso{
\code{\link[base]{print}}, \code{\link[utils]{capture.output}}
}

