% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assert-r-has-capability.R, R/r-has-capability.R
\name{assert_r_can_build_translations}
\alias{assert_r_can_build_translations}
\alias{r_can_build_translations}
\title{Can R find tools for building translations?}
\usage{
assert_r_can_build_translations()

r_can_build_translations()
}
\value{
\code{r_can_build_translations} returns \code{TRUE} if R can see
\code{gettext} and \code{msgfmt} tools, and \code{FALSE} (with a cause)
otherwise.
\code{assert_r_can_build_translations} returns nothing but throws an error if
\code{r_can_build_translations} function returns \code{FALSE}.
}
\description{
Checks to see if R can see the \code{gettext} and \code{msgfmt} tools in
order to build binary translation files.
}
\examples{
r_can_build_translations()
dont_stop(assert_r_can_build_translations())
}
\seealso{
\code{\link{r_can_compile_code}}
}

