% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assert-r-has-capability.R, R/r-has-capability.R
\name{assert_r_can_compile_code}
\alias{assert_r_can_compile_code}
\alias{r_can_compile_code}
\title{Can R find tools for compiling code?}
\usage{
assert_r_can_compile_code()

r_can_compile_code()
}
\value{
\code{r_can_compile_code} returns \code{TRUE} if R can see \code{gcc}
and \code{make} tools, and \code{FALSE} (with a cause) otherwise.
\code{assert_r_can_compile_code} returns nothing but throws an error if
\code{r_can_compile_code} function returns \code{FALSE}.
}
\description{
Checks to see if R can see the \code{gcc} and \code{make} tools in order
to compile code.
}
\examples{
r_can_compile_code()
dont_stop(assert_r_can_compile_code())
}
\seealso{
\code{\link{r_can_build_translations}}
}

