% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{r_has_jpeg_capability}
\alias{r_has_jpeg_capability}
\alias{r_has_png_capability}
\alias{r_has_tiff_capability}
\alias{r_has_tcltk_capability}
\alias{r_has_x11_capability}
\alias{r_has_aqua_capability}
\alias{r_has_http_ftp_capability}
\alias{r_has_sockets_capability}
\alias{r_has_libxml_capability}
\alias{r_has_fifo_capability}
\alias{r_has_cledit_capability}
\alias{r_has_iconv_capability}
\alias{r_has_nls_capability}
\alias{r_has_rprof_capability}
\alias{r_has_profmem_capability}
\alias{r_has_cairo_capability}
\alias{r_has_icu_capability}
\alias{r_has_long_double_capability}
\alias{r_has_libcurl_capability}
\alias{assert_r_has_jpeg_capability}
\alias{assert_r_has_png_capability}
\alias{assert_r_has_tiff_capability}
\alias{assert_r_has_tcltk_capability}
\alias{assert_r_has_x11_capability}
\alias{assert_r_has_aqua_capability}
\alias{assert_r_has_http_ftp_capability}
\alias{assert_r_has_sockets_capability}
\alias{assert_r_has_libxml_capability}
\alias{assert_r_has_fifo_capability}
\alias{assert_r_has_cledit_capability}
\alias{assert_r_has_iconv_capability}
\alias{assert_r_has_nls_capability}
\alias{assert_r_has_rprof_capability}
\alias{assert_r_has_profmem_capability}
\alias{assert_r_has_cairo_capability}
\alias{assert_r_has_icu_capability}
\alias{assert_r_has_long_double_capability}
\alias{assert_r_has_libcurl_capability}
\title{Does R have a capability?}
\description{
See \code{\link[assertive.reflection]{r_has_jpeg_capability}}.
}
