% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{set_cause}
\alias{set_cause}
\title{Set a cause and return the input}
\usage{
set_cause(x, value)
}
\arguments{
\item{x}{A variable.}

\item{value}{A character vector to set the cause to, where \code{x} is
not \code{TRUE}.}
}
\value{
\code{x}, with a new cause attribute.
}
\description{
Sets the cause attribute of an object and returns that object.
}
\details{
If \code{x} is \code{TRUE} everywhere, this returns the input
without setting a cause.  Otherwise, the cause is an empty string where
\code{x} is \code{TRUE}, and \code{value} elsewhere.
}
\seealso{
\code{\link{cause}} , \code{\link[stats]{setNames}}
}

