% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{truncate}
\alias{truncate}
\title{Truncate a string}
\usage{
truncate(x, width = getOption("width"))
}
\arguments{
\item{x}{A character vector, or something coercible to one.}

\item{width}{A positive integer.}
}
\value{
A character vector
}
\description{
Truncates a character vector to have a maximum length.
}
\examples{
\dontrun{
truncate(c("abcd", "efghi", "jklmno", "pqrstuv"), 5)
}
}

