% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-base.R
\name{Truth}
\alias{Truth}
\alias{assert_all_are_false}
\alias{assert_all_are_na}
\alias{assert_all_are_not_false}
\alias{assert_all_are_not_na}
\alias{assert_all_are_not_true}
\alias{assert_all_are_true}
\alias{assert_any_are_false}
\alias{assert_any_are_na}
\alias{assert_any_are_not_false}
\alias{assert_any_are_not_na}
\alias{assert_any_are_not_true}
\alias{assert_any_are_true}
\alias{assert_is_identical_to_false}
\alias{assert_is_identical_to_na}
\alias{assert_is_identical_to_true}
\alias{is_false}
\alias{is_identical_to_false}
\alias{is_identical_to_na}
\alias{is_identical_to_true}
\alias{is_na}
\alias{is_not_false}
\alias{is_not_na}
\alias{is_not_true}
\alias{is_true}
\title{Is the input TRUE/FALSE/NA?}
\description{
See \code{\link[assertive.base]{Truth}}.
}

