% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-type.R
\name{type_description}
\alias{type_description}
\title{Describe the type of object}
\usage{
type_description(x)
}
\arguments{
\item{x}{A variable.}
}
\value{
A string.
}
\description{
Get the class or mode (for arrays).
}

