% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-file.R
\name{warn_about_file.access_under_windows}
\alias{warn_about_file.access_under_windows}
\title{Warn about file.access under Windows}
\usage{
warn_about_file.access_under_windows()
}
\value{
Nothing. Invoked for the side effect of throwing a warning under
Windows.
}
\description{
If the OS is Windows, throw a warning about using
\code{\link[base]{file.access}}.
}
\examples{
\dontrun{
dont_stop(warn_about_file.access_under_windows())
}
}
\seealso{
\code{\link[base]{file.access}}
}

