% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{bapply}
\alias{bapply}
\title{Wrapper to vapply that returns booleans.}
\usage{
bapply(x, predicate, ...)
}
\arguments{
\item{x}{A vector (atomic or list).}

\item{predicate}{A predicate (function that returns a bool) to apply.
elementwise to \code{x}.}

\item{...}{Passed to \code{vapply}.}
}
\value{
A logical vector.
}
\description{
Wrapper to \code{\link{vapply}} for functions that return a boolean (logical
scalar) value.
}
\note{
\code{USE.NAMES} is set to \code{TRUE}
}
\seealso{
\code{\link{vapply}}.
}

