% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-base.R, R/exports-properties.R
\name{DIM}
\alias{DIM}
\title{Get the dimensions of an object}
\description{
See \code{\link[assertive.properties]{DIM}}.

See \code{\link[assertive.properties]{DIM}}.
}

