% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{has_arg}
\alias{assert_has_arg}
\alias{has_arg}
\alias{has_arg_}
\title{Does the current call have an argument?
 
See \code{\link[assertive.code]{has_arg}}.}
\description{
Does the current call have an argument?
 
See \code{\link[assertive.code]{has_arg}}.
}

