% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{has_cols}
\alias{assert_has_cols}
\alias{assert_has_rows}
\alias{has_cols}
\alias{has_rows}
\title{Does the input have rows/columns?}
\description{
See \code{\link[assertive.properties]{has_cols}}.
}

