% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{has_duplicates}
\alias{has_duplicates}
\alias{has_no_duplicates}
\alias{assert_has_duplicates}
\alias{assert_has_no_duplicates}
\title{Does the input have duplicates?}
\description{
See \code{\link[assertive.properties]{has_duplicates}}.
}
