% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{has_names}
\alias{assert_has_colnames}
\alias{assert_has_dimnames}
\alias{assert_has_names}
\alias{assert_has_rownames}
\alias{has_colnames}
\alias{has_dimnames}
\alias{has_names}
\alias{has_rownames}
\title{Does the input have names?}
\description{
See \code{\link[assertive.properties]{has_names}}.
}

