% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_array}
\alias{assert_is_array}
\alias{assert_is_matrix}
\alias{is_array}
\alias{is_matrix}
\title{Is the input an array or matrix?}
\description{
See \code{\link[assertive.types]{is_array}}.
}

