% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_binding_locked}
\alias{assert_is_binding_locked}
\alias{is_binding_locked}
\title{Is the binding of a variable locked?}
\description{
See \code{\link[assertive.code]{is_binding_locked}}.
}

