% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_character}
\alias{assert_is_a_string}
\alias{assert_is_character}
\alias{is_a_string}
\alias{is_character}
\title{Is the input of type character?}
\description{
See \code{\link[assertive.types]{is_character}}.
}

