% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_class}
\alias{is_class}
\alias{assert_all_are_classes}
\alias{assert_any_are_classes}
\title{Is the input the name of a (formally defined) class?}
\description{
See \code{\link[assertive.types]{is_class}}.
}
