% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-files.R
\name{is_connection}
\alias{assert_is_bzfile_connection}
\alias{assert_is_connection}
\alias{assert_is_fifo_connection}
\alias{assert_is_file_connection}
\alias{assert_is_gzfile_connection}
\alias{assert_is_incomplete_connection}
\alias{assert_is_open_connection}
\alias{assert_is_pipe_connection}
\alias{assert_is_readable_connection}
\alias{assert_is_socket_connection}
\alias{assert_is_stderr}
\alias{assert_is_stdin}
\alias{assert_is_stdout}
\alias{assert_is_terminal_connection}
\alias{assert_is_text_connection}
\alias{assert_is_unz_connection}
\alias{assert_is_url_connection}
\alias{assert_is_writable_connection}
\alias{assert_is_xzfile_connection}
\alias{is_bzfile_connection}
\alias{is_connection}
\alias{is_fifo_connection}
\alias{is_file_connection}
\alias{is_gzfile_connection}
\alias{is_incomplete_connection}
\alias{is_open_connection}
\alias{is_pipe_connection}
\alias{is_readable_connection}
\alias{is_socket_connection}
\alias{is_stderr}
\alias{is_stdin}
\alias{is_stdout}
\alias{is_terminal_connection}
\alias{is_text_connection}
\alias{is_unz_connection}
\alias{is_url_connection}
\alias{is_writable_connection}
\alias{is_xzfile_connection}
\title{Is the input a connection?}
\description{
See \code{\link[assertive.files]{is_connection}}.
}

