% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_credit_card_number}
\alias{assert_all_are_credit_card_numbers}
\alias{assert_any_are_credit_card_numbers}
\alias{is_credit_card_number}
\title{Does the character vector contain credit card numbers?}
\description{
See \code{\link[assertive.data]{is_credit_card_number}}.
}

