% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-datetimes.R
\name{is_date_string}
\alias{assert_all_are_date_strings}
\alias{assert_any_are_date_strings}
\alias{is_date_string}
\title{Does the character vector contain dates?}
\description{
See \code{\link[assertive.datetimes]{is_date_string}}.
}

