% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-numbers.R
\name{is_divisible_by}
\alias{assert_all_are_divisible_by}
\alias{assert_all_are_even}
\alias{assert_all_are_odd}
\alias{assert_any_are_divisible_by}
\alias{assert_any_are_even}
\alias{assert_any_are_odd}
\alias{is_divisible_by}
\alias{is_even}
\alias{is_odd}
\title{Is the input divisible by a number?}
\description{
See \code{\link[assertive.numbers]{is_divisible_by}}.
}

