% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-files.R
\name{is_empty_file}
\alias{assert_all_are_empty_files}
\alias{assert_all_are_non_empty_files}
\alias{assert_any_are_empty_files}
\alias{assert_any_are_non_empty_files}
\alias{is_empty_file}
\alias{is_file_size_in_range}
\alias{is_non_empty_file}
\title{Is a file empty?}
\description{
See \code{\link[assertive.files]{is_empty_file}}.
}

