% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-models.R
\name{is_empty_model}
\alias{assert_is_empty_model}
\alias{assert_is_non_empty_model}
\alias{is_empty_model}
\alias{is_non_empty_model}
\title{Is the input the empty model?}
\description{
See \code{\link[assertive.models]{is_empty_model}}.
}

