% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assert-is-file.R, R/is-file.R
\name{assert_all_are_executable_files}
\alias{assert_all_are_executable_files}
\alias{assert_all_are_readable_files}
\alias{assert_all_are_writable_files}
\alias{assert_any_are_executable_files}
\alias{assert_any_are_readable_files}
\alias{assert_any_are_writable_files}
\alias{is_ex_file}
\alias{is_executable_file}
\alias{is_readable_file}
\alias{is_writable_file}
\title{Is the file accessible?}
\usage{
assert_all_are_executable_files(x)

assert_any_are_executable_files(x)

assert_all_are_readable_files(x)

assert_any_are_readable_files(x)

assert_all_are_writable_files(x)

assert_any_are_writable_files(x)

is_executable_file(x)

is_ex_file(x)

is_readable_file(x)

is_writable_file(x)
}
\arguments{
\item{x}{Input to check.}
}
\value{
\code{is_executable_file} wraps \code{file.access}, showing
the names of the inputs in the answer.   \code{assert_is_executable_file}
returns nothing but throws an error if \code{is_executable_file} returns
\code{FALSE}.
}
\description{
Checks to see if the input files can be executed/read/written to.
}
\examples{
\dontrun{
assert_all_are_readable_files(dir())
}
}
\seealso{
\code{\link[base]{file.access}}.
}

