% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_existing}
\alias{is_existing}
\alias{assert_all_are_existing}
\alias{assert_any_are_existing}
\title{Does the variable exist?
 
See \code{\link[assertive.code]{is_existing}}.}
\description{
Does the variable exist?
 
See \code{\link[assertive.code]{is_existing}}.
}
