% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-numbers.R
\name{is_finite}
\alias{assert_all_are_finite}
\alias{assert_all_are_infinite}
\alias{assert_all_are_negative_infinity}
\alias{assert_all_are_positive_infinity}
\alias{assert_any_are_finite}
\alias{assert_any_are_infinite}
\alias{assert_any_are_negative_infinity}
\alias{assert_any_are_positive_infinity}
\alias{is_finite}
\alias{is_infinite}
\alias{is_negative_infinity}
\alias{is_positive_infinity}
\title{Are the inputs (in)finite?}
\description{
See \code{\link[assertive.numbers]{is_finite}}.
}

