% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_function}
\alias{assert_is_function}
\alias{assert_is_primitive}
\alias{assert_is_stepfun}
\alias{is_function}
\alias{is_primitive}
\alias{is_stepfun}
\title{Is the input a function?}
\description{
See \code{\link[assertive.types]{is_function}}.
}

