% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_honorific}
\alias{assert_all_are_honorifics}
\alias{assert_any_are_honorifics}
\alias{is_honorific}
\title{Is the string an honorific?}
\description{
See \code{\link[assertive.data]{is_honorific}}.
}

