% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_if_condition}
\alias{assert_is_if_condition}
\alias{is_if_condition}
\title{Is suitable to be used as an if condition
 
See \code{\link[assertive.code]{is_if_condition}}.}
\description{
Is suitable to be used as an if condition
 
See \code{\link[assertive.code]{is_if_condition}}.
}

