% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-datetimes.R
\name{is_in_past}
\alias{assert_all_are_after}
\alias{assert_all_are_before}
\alias{assert_all_are_in_future}
\alias{assert_all_are_in_past}
\alias{assert_any_are_after}
\alias{assert_any_are_before}
\alias{assert_any_are_in_future}
\alias{assert_any_are_in_past}
\alias{is_after}
\alias{is_before}
\alias{is_in_future}
\alias{is_in_past}
\title{Is the input in the past/future?}
\description{
See \code{\link[assertive.datetimes]{is_in_past}}.
}

