% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_integer}
\alias{assert_is_an_integer}
\alias{assert_is_integer}
\alias{is_an_integer}
\alias{is_integer}
\title{Is the input an integer?}
\description{
See \code{\link[assertive.types]{is_integer}}.
}

