% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data.R
\name{is_ip_address}
\alias{assert_all_are_ip_addresses}
\alias{assert_any_are_ip_addresses}
\alias{is_ip_address}
\title{Does the character vector contain IP addresses?}
\description{
See \code{\link[assertive.data]{is_ip_address}}.
}

