% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_language}
\alias{assert_is_call}
\alias{assert_is_expression}
\alias{assert_is_language}
\alias{assert_is_name}
\alias{assert_is_symbol}
\alias{is_call}
\alias{is_expression}
\alias{is_language}
\alias{is_name}
\alias{is_symbol}
\title{Is the input a language object?}
\description{
See \code{\link[assertive.types]{is_language}}.
}

