% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-files.R
\name{is_library}
\alias{assert_all_are_libraries}
\alias{assert_any_are_libraries}
\alias{is_library}
\title{Is the directory a known R library?}
\description{
See \code{\link[assertive.files]{is_library}}.
}

