% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_loaded}
\alias{is_loaded}
\alias{assert_is_loaded}
\title{Is the input DLL loaded?
 
See \code{\link[assertive.code]{is_loaded}}.}
\description{
Is the input DLL loaded?
 
See \code{\link[assertive.code]{is_loaded}}.
}
