% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-properties.R
\name{is_null}
\alias{assert_is_not_null}
\alias{assert_is_null}
\alias{is_not_null}
\alias{is_null}
\title{Checks to see if the input is (not) null.}
\description{
See \code{\link[assertive.properties]{is_null}}.
}

