% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-strings.R
\name{is_numeric_string}
\alias{is_numeric_string}
\alias{is_logical_string}
\alias{assert_all_are_numeric_strings}
\alias{assert_any_are_numeric_strings}
\alias{assert_all_are_logical_strings}
\alias{assert_any_are_logical_strings}
\title{Does the string contain a number?
 
See \code{\link[assertive.strings]{is_numeric_string}}.}
\description{
Does the string contain a number?
 
See \code{\link[assertive.strings]{is_numeric_string}}.
}
