% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_on_os_path}
\alias{assert_all_are_on_os_path}
\alias{assert_any_are_on_os_path}
\alias{is_on_os_path}
\title{Is the path on the OS path?}
\description{
See \code{\link[assertive.reflection]{is_on_os_path}}.
}

