% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_package_current}
\alias{assert_all_are_current_packages}
\alias{assert_any_are_current_packages}
\alias{assert_is_package_current}
\alias{is_package_current}
\title{Is the installed version of a package current?}
\description{
See \code{\link[assertive.reflection]{is_package_current}}.
}

