% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_r}
\alias{assert_is_architect}
\alias{assert_is_r}
\alias{assert_is_r_alpha}
\alias{assert_is_r_beta}
\alias{assert_is_r_devel}
\alias{assert_is_r_patched}
\alias{assert_is_r_release}
\alias{assert_is_r_release_candidate}
\alias{assert_is_r_stable}
\alias{assert_is_revo_r}
\alias{assert_is_rstudio}
\alias{is_architect}
\alias{is_r}
\alias{is_r_alpha}
\alias{is_r_beta}
\alias{is_r_devel}
\alias{is_r_patched}
\alias{is_r_release}
\alias{is_r_release_candidate}
\alias{is_r_stable}
\alias{is_revo_r}
\alias{is_rstudio}
\title{Are you running R?}
\description{
See \code{\link[assertive.reflection]{is_r}}.
}

