% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assert-is-reflection.R, R/is-reflection.R
\name{assert_is_architect}
\alias{assert_is_architect}
\alias{assert_is_r}
\alias{assert_is_r_devel}
\alias{assert_is_r_patched}
\alias{assert_is_r_stable}
\alias{assert_is_revo_r}
\alias{assert_is_rstudio}
\alias{assert_is_slave_r}
\alias{is_architect}
\alias{is_r}
\alias{is_r_devel}
\alias{is_r_patched}
\alias{is_r_stable}
\alias{is_revo_r}
\alias{is_rstudio}
\alias{is_slave_r}
\title{Are you running R?}
\usage{
assert_is_architect()

assert_is_r()

assert_is_r_devel()

assert_is_r_patched()

assert_is_r_stable()

assert_is_revo_r()

assert_is_rstudio()

assert_is_slave_r()

is_architect()

is_r()

is_r_devel()

is_r_patched()

is_r_stable()

is_revo_r()

is_rstudio()

is_slave_r()
}
\value{
\code{is_r} wraps \code{is.R}, providing more
information on failure.  \code{is_r_stable}, \code{is_r_patched} and
\code{is_r_devel} tell you what type of R build you are
running.  \code{is_architect}, \code{is_rstudio} and \code{is_revo_r} tell
you if you are running Architect/StatET, RStudio, or Revolution Analytics'
Revolution R build.  \code{is_slave_r} tells you if you are running a slave
instance of R (e.g. when building a package with \code{devtools} or using a
cluster).
The \code{assert_*} functions return nothing but throw an error if
the corresponding \code{is_*} function returns \code{FALSE}.
}
\description{
Checks to see what type of R you are running.
}
\examples{
is_r()
is_r_stable()
is_r_patched()
is_r_devel()
is_architect()
is_revo_r()
is_rstudio()
is_slave_r()
switch(
  version$status,
  "Patched"                      = assert_is_r_patched(),
  "Under development (unstable)" = assert_is_r_devel(),
  assert_is_r_stable()
)
dont_stop(assert_is_r())
}
\references{
\url{http://www.revolutionanalytics.com/revolution-r-open}
}
\seealso{
\code{\link[base]{is.R}}, \code{\link[base]{version}}.
}

