% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-numbers.R
\name{is_real}
\alias{assert_all_are_imaginary}
\alias{assert_all_are_real}
\alias{assert_any_are_imaginary}
\alias{assert_any_are_real}
\alias{is_imaginary}
\alias{is_real}
\title{Is the input real/imaginary?}
\description{
See \code{\link[assertive.numbers]{is_real}}.
}

