% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-reflection.R
\name{is_rstudio_current}
\alias{is_rstudio_current}
\title{Is RStudio the current version?}
\usage{
is_rstudio_current()
}
\value{
\code{is_rstudio_current} returns \code{TRUE} or \code{FALSE}, and
\code{assert_is_rstudio_current} throws an error in the event of an out of
date RStudio.  Non-RStudio IDEs throw an error.
}
\description{
Checks to see if the running version of RStudio is the current version.
}
\references{
This function is engineered from the \code{downloadUpdateInfo}
function from
\url{https://github.com/rstudio/rstudio/blob/master/src/cpp/session/modules/SessionUpdates.R}
where the string for the OS is described in \code{beginUpdateCheck} from
\url{https://github.com/rstudio/rstudio/blob/master/src/cpp/session/modules/SessionUpdates.cpp}
}
\seealso{
\code{\link{is_rstudio}}, \code{\link{is_rstudio_desktop}}
}

