% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-matrices.R
\name{is_symmetric_matrix}
\alias{assert_is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Is the input a symmetric matrix?}
\description{
See \code{\link[assertive.matrices]{is_symmetric_matrix}}.
}

