% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-types.R
\name{is_ts}
\alias{assert_is_mts}
\alias{assert_is_ts}
\alias{assert_is_tskernel}
\alias{is_mts}
\alias{is_ts}
\alias{is_tskernel}
\title{Is the input a time series?}
\description{
See \code{\link[assertive.types]{is_ts}}.
}

