% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data-uk.R
\name{is_uk_car_licence}
\alias{is_uk_car_licence}
\alias{is_uk_car_license}
\alias{assert_all_are_uk_car_licences}
\alias{assert_any_are_uk_car_licences}
\alias{assert_all_are_uk_car_licenses}
\alias{assert_any_are_uk_car_licenses}
\title{Is the string a valid UK car licence plate number?}
\description{
See \code{\link[assertive.data.uk]{is_uk_car_licence}}.
}
