% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data-uk.R
\name{is_uk_national_insurance_number}
\alias{assert_all_are_uk_national_insurance_numbers}
\alias{assert_any_are_uk_national_insurance_numbers}
\alias{is_uk_national_insurance_number}
\title{Is the string a valid UK national insurance number?}
\description{
See \code{\link[assertive.data.uk]{is_uk_national_insurance_number}}.
}

