% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data-uk.R
\name{is_uk_postcode}
\alias{assert_all_are_uk_postcodes}
\alias{assert_any_are_uk_postcodes}
\alias{is_uk_postcode}
\title{Is the string a valid UK postcode?}
\description{
See \code{\link[assertive.data.uk]{is_uk_postcode}}.
}

