% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data-uk.R
\name{is_uk_telephone_number}
\alias{assert_all_are_uk_telephone_numbers}
\alias{assert_any_are_uk_telephone_numbers}
\alias{is_uk_telephone_number}
\title{Is the string a valid UK telephone number?}
\description{
See \code{\link[assertive.data.uk]{is_uk_telephone_number}}.
}

