% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-data-us.R
\name{is_us_telephone_number}
\alias{is_us_telephone_number}
\alias{assert_all_are_us_telephone_numbers}
\alias{assert_any_are_us_telephone_numbers}
\title{Is the string a valid US telephone number?}
\description{
See \code{\link[assertive.data.us]{is_us_telephone_number}}.
}
