% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-data-us.R
\name{is_us_zip_code}
\alias{assert_all_are_us_zip_codes}
\alias{assert_any_are_us_zip_codes}
\alias{is_us_zip_code}
\title{Is the string a valid US zip code?}
\description{
See \code{\link[assertive.data.us]{is_us_zip_code}}.
}

