% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_valid_r_code}
\alias{is_valid_r_code}
\alias{assert_is_valid_r_code}
\title{Is the input valid R code?
 
See \code{\link[assertive.code]{is_valid_r_code}}.}
\description{
Is the input valid R code?
 
See \code{\link[assertive.code]{is_valid_r_code}}.
}
