% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-code.R
\name{is_valid_variable_name}
\alias{assert_all_are_valid_variable_names}
\alias{assert_any_are_valid_variable_names}
\alias{is_valid_variable_name}
\title{Is the string a valid variable name?}
\description{
See \code{\link[assertive.code]{is_valid_variable_name}}.
}

