% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-numbers.R
\name{is_whole_number}
\alias{assert_all_are_whole_numbers}
\alias{assert_all_numbers_are_whole_numbers}
\alias{assert_any_are_whole_numbers}
\alias{assert_any_numbers_are_whole_numbers}
\alias{is_whole_number}
\title{Is the input a whole number?}
\description{
See \code{\link[assertive.numbers]{is_whole_number}}.
}

