% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{is_windows}
\alias{assert_is_32_bit}
\alias{assert_is_32_bit_os}
\alias{assert_is_64_bit}
\alias{assert_is_64_bit_os}
\alias{assert_is_bsd}
\alias{assert_is_linux}
\alias{assert_is_mac}
\alias{assert_is_macos_sierra}
\alias{assert_is_osx}
\alias{assert_is_osx_cheetah}
\alias{assert_is_osx_el_capitan}
\alias{assert_is_osx_jaguar}
\alias{assert_is_osx_leopard}
\alias{assert_is_osx_lion}
\alias{assert_is_osx_mavericks}
\alias{assert_is_osx_mountain_lion}
\alias{assert_is_osx_panther}
\alias{assert_is_osx_puma}
\alias{assert_is_osx_snow_leopard}
\alias{assert_is_osx_tiger}
\alias{assert_is_osx_yosemite}
\alias{assert_is_solaris}
\alias{assert_is_unix}
\alias{assert_is_windows}
\alias{assert_is_windows_10}
\alias{assert_is_windows_7}
\alias{assert_is_windows_8}
\alias{assert_is_windows_8.1}
\alias{assert_is_windows_server_2008}
\alias{assert_is_windows_server_2008_r2}
\alias{assert_is_windows_server_2012}
\alias{assert_is_windows_server_2012_r2}
\alias{assert_is_windows_vista}
\alias{is_32_bit}
\alias{is_32_bit_os}
\alias{is_64_bit}
\alias{is_64_bit_os}
\alias{is_bsd}
\alias{is_linux}
\alias{is_mac}
\alias{is_macos_sierra}
\alias{is_osx}
\alias{is_osx_cheetah}
\alias{is_osx_el_capitan}
\alias{is_osx_jaguar}
\alias{is_osx_leopard}
\alias{is_osx_lion}
\alias{is_osx_mavericks}
\alias{is_osx_mountain_lion}
\alias{is_osx_panther}
\alias{is_osx_puma}
\alias{is_osx_snow_leopard}
\alias{is_osx_tiger}
\alias{is_osx_yosemite}
\alias{is_solaris}
\alias{is_unix}
\alias{is_windows}
\alias{is_windows_10}
\alias{is_windows_7}
\alias{is_windows_8}
\alias{is_windows_8.1}
\alias{is_windows_server_2008}
\alias{is_windows_server_2008_r2}
\alias{is_windows_server_2012}
\alias{is_windows_server_2012_r2}
\alias{is_windows_vista}
\title{What OS is running?}
\description{
See \code{\link[assertive.reflection]{is_windows}}.
}

