% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal-string.R
\name{matches_regex}
\alias{matches_regex}
\title{Does the input match the regular expression?}
\usage{
matches_regex(x, rx, ignore.case = TRUE, ...)
}
\arguments{
\item{x}{Input to check.}

\item{rx}{A regular expression.}

\item{ignore.case}{Should the case of alphabetic characters be ignored?}

\item{...}{Passed to \code{\link{grepl}}.}
}
\value{
A logical vector that is \code{TRUE} when the input matches the
regular expression.
}
\description{
Checks that the input matches the regular expression.
}
\note{
The default for \code{ignore.case} is different to the default in
\code{grepl}.
}
\seealso{
\code{\link{regex}} and \code{\link{regexpr}}.
}

