% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{na}
\alias{na}
\title{NA, with a cause of failure.}
\usage{
na(...)
}
\arguments{
\item{...}{Passed to \code{gettextf} to create a cause of failure message.}
}
\value{
\code{NA} with the attribute \code{cause}, as provided
in the input.
}
\description{
Always returns the value (logical) \code{NA}, with a cause attribute.
}
\seealso{
\code{\link{cause}} and \code{\link{false}}.
}

