% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exports-reflection.R
\name{r_can_find_tools}
\alias{assert_r_can_build_translations}
\alias{assert_r_can_compile_code}
\alias{assert_r_can_find_tools}
\alias{r_can_build_translations}
\alias{r_can_compile_code}
\alias{r_can_find_tools}
\title{Can R find tools?}
\description{
See \code{\link[assertive.reflection]{r_can_find_tools}}.
}

